small updates

+++ model.py
@@ -8,10 +8,10 @@ class BasicCNN(nn.Module):
         self.features = nn.Sequential(
             # in_channels: number of channels, for CT/MRI = 1, RGB = 3; out_channels: how many filters to apply
             # both independent of image dimensions (width, height, depth)
-            nn.Conv3d(1, 16, kernel_size=3, stride=1, padding=1).to(torch.float32),
+            nn.Conv3d(1, 16, kernel_size=3, stride=1, padding=1),
             nn.ReLU(inplace=True),
             nn.MaxPool3d(kernel_size=2, stride=2),
-            nn.Conv3d(16, 32, kernel_size=3, stride=1, padding=1).to(torch.float32),
+            nn.Conv3d(16, 32, kernel_size=3, stride=1, padding=1),
             nn.ReLU(inplace=True),
             nn.MaxPool3d(kernel_size=2, stride=2)
         )
